move Dashboard to pages and fix refrence
