+++ ClientReact/src/App.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import { Center } from '@mantine/core'
 import SignIn from './pages/SignIn'
 import Landing from './pages/Landing'
-import Dashboard from './components/Dashboard'
+import Dashboard from './pages/Dashboard'
 import Credits from './pages/Credits'
 import './App.css'
